Rename scheduled tasks module
