+++ backend/scheduled/activate_trials.py
